+++ src/components/MaterialTheme/TwentyFourHoursMode.jsx
@@ -71,15 +71,17 @@ class TwentyFourHoursMode extends React.PureComponent {
   }
 
   setStep(step) {
-    const pointerRotate = step === 0 ? this.resetHourDegree() : this.resetMinuteDegree();
+    const pointerRotate = step === 0
+      ? this.resetHourDegree()
+      : this.resetMinuteDegree();
     this.setState({
       step,
       pointerRotate
     });
   }
 
-  handleTimePointerClick(time, pointerRotate) {
-    this.setState({pointerRotate});
+  handleTimePointerClick(time, pointerRotate = null) {
+    pointerRotate && this.setState({pointerRotate});
     this.handleTimeChange(time);
   }
 
@@ -111,7 +113,9 @@ class TwentyFourHoursMode extends React.PureComponent {
     let pointerRotate = 0;
     HOURS.forEach((h, index) => {
       if (hour === index + 1) {
-        pointerRotate = index < 12 ? 360 * (index + 1) / 12 : 360 * (index + 1 - 12) / 12;
+        pointerRotate = index < 12
+          ? 360 * (index + 1) / 12
+          : 360 * (index + 1 - 12) / 12;
       }
     });
     return pointerRotate;
@@ -133,7 +137,9 @@ class TwentyFourHoursMode extends React.PureComponent {
     let {hour, minute} = this.props;
     let time = step === 0 ? hour : minute;
     let splitNum = step === 0 ? 12 : 60;
-    let minLength = step === 0 ? MIN_ABSOLUTE_POSITION : MAX_ABSOLUTE_POSITION;
+    let minLength = step === 0
+      ? MIN_ABSOLUTE_POSITION
+      : MAX_ABSOLUTE_POSITION;
     let height = time < splitNum
       ? minLength - POINTER_RADIUS
       : MAX_ABSOLUTE_POSITION - POINTER_RADIUS;
@@ -147,12 +153,14 @@ class TwentyFourHoursMode extends React.PureComponent {
     const {
       hour,
       minute,
-      draggable,
       phrases,
-      showTimezone,
       timezone,
+      draggable,
+      limitDrag,
+      minuteStep,
+      showTimezone,
       timezoneIsEditable,
-      onTimezoneChange
+      onTimezoneChange,
     } = this.props;
 
     const {step, pointerRotate} = this.state;
@@ -177,24 +185,36 @@ class TwentyFourHoursMode extends React.PureComponent {
         <div className='time_picker_modal_header'>
           <span
             className={activeHourClass}
-            onClick={this.handleStepChange.bind(this, 0)}>{hour}</span>
+            onClick={this.handleStepChange.bind(this, 0)}
+          >
+            {hour}
+          </span>
           <span className='time_picker_header_delivery'>:</span>
-          <span className={activeMinuteClass}
-            onClick={this.handleStepChange.bind(this, 1)}>{minute}</span>
+          <span
+            className={activeMinuteClass}
+            onClick={this.handleStepChange.bind(this, 1)}
+          >
+            {minute}
+          </span>
         </div>
         <div className='picker_container'>
           <PickerPointGenerator
             ref={ref => this.pickerPointerContainer = ref}
             handleTimePointerClick={this.handleTimePointerClick}
-            rotateState={rotateState}
+            pointerRotate={pointerRotate}
           />
           <PickerDragHandler
             step={step}
+            limitDrag={limitDrag}
+            minuteStep={minuteStep}
             draggable={draggable}
             rotateState={rotateState}
             time={step === 0 ? parseInt(hour) : parseInt(minute)}
-            minLength={step === 0 ? MIN_ABSOLUTE_POSITION : MAX_ABSOLUTE_POSITION}
-            handleTimePointerClick={this.handleTimePointerClick} />
+            minLength={step === 0
+                ? MIN_ABSOLUTE_POSITION
+                : MAX_ABSOLUTE_POSITION}
+            handleTimePointerClick={this.handleTimePointerClick}
+          />
         </div>
         {showTimezone
           ? <Timezone