+++ stories/TimePicker.js
@@ -1,10 +1,10 @@
 import '../css/material/default.css';
 
-import { text, withKnobs } from '@kadira/storybook-addon-knobs';
+import {text, withKnobs} from '@kadira/storybook-addon-knobs';
 
 import React from 'react';
 import TimePickerWrapper from '../examples/TimePickerWrapper';
-import { storiesOf } from '@kadira/storybook';
+import {storiesOf} from '@kadira/storybook';
 
 storiesOf('Default TimePicker', module)
   .addDecorator(withKnobs)
@@ -27,5 +27,18 @@ storiesOf('Default TimePicker', module)
   .addWithInfo('undraggable', () => (
     <TimePickerWrapper
       draggable={false}
+    />
+  ))
+  .addWithInfo('custom minute step', () => (
+    <TimePickerWrapper
+      autoMode={false}
+      minuteStep={1}
+    />
+  ))
+  .addWithInfo('limit drag', () => (
+    <TimePickerWrapper
+      limitDrag
+      autoMode={false}
+      minuteStep={1}
     />
   ));