+++ stories/TwelveHoursMode.js
@@ -2,7 +2,7 @@ import '../css/material/default.css';
 
 import React from 'react';
 import TimePickerWrapper from '../examples/TimePickerWrapper';
-import { storiesOf } from '@kadira/storybook';
+import {storiesOf} from '@kadira/storybook';
 
 storiesOf('TwelveHoursMode', module)
   .addWithInfo('with default time', () => (
@@ -21,5 +21,20 @@ storiesOf('TwelveHoursMode', module)
       withoutIcon={true}
       timeMode="12"
       focused={true}
+    />
+  ))
+  .addWithInfo('custom minute step', () => (
+    <TimePickerWrapper
+      autoMode={false}
+      minuteStep={1}
+      timeMode="12"
+    />
+  ))
+  .addWithInfo('limit drag', () => (
+    <TimePickerWrapper
+      limitDrag
+      autoMode={false}
+      minuteStep={1}
+      timeMode="12"
     />
   ));